+++ Sources/Requests/Requests+CargoByDriverSharingSettingsUpdateReq.swift
@@ -4,7 +4,6 @@ extension Requests {
     public struct CargoByDriverSharingSettingsUpdateReq: Codable, Hashable, Sendable {
         public let taskGroupId: String
         public let cargoId: String
-        public let driverId: String
         public let `public`: Bool
         /// Additional properties that are not explicitly defined in the schema
         public let additionalProperties: [String: JSONValue]
@@ -12,13 +11,11 @@ extension Requests {
         public init(
             taskGroupId: String,
             cargoId: String,
-            driverId: String,
             public: Bool,
             additionalProperties: [String: JSONValue] = .init()
         ) {
             self.taskGroupId = taskGroupId
             self.cargoId = cargoId
-            self.driverId = driverId
             self.public = `public`
             self.additionalProperties = additionalProperties
         }
@@ -27,7 +24,6 @@ extension Requests {
             let container = try decoder.container(keyedBy: CodingKeys.self)
             self.taskGroupId = try container.decode(String.self, forKey: .taskGroupId)
             self.cargoId = try container.decode(String.self, forKey: .cargoId)
-            self.driverId = try container.decode(String.self, forKey: .driverId)
             self.public = try container.decode(Bool.self, forKey: .public)
             self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
         }
@@ -37,7 +33,6 @@ extension Requests {
             try encoder.encodeAdditionalProperties(self.additionalProperties)
             try container.encode(self.taskGroupId, forKey: .taskGroupId)
             try container.encode(self.cargoId, forKey: .cargoId)
-            try container.encode(self.driverId, forKey: .driverId)
             try container.encode(self.public, forKey: .public)
         }
 
@@ -45,7 +40,6 @@ extension Requests {
         enum CodingKeys: String, CodingKey, CaseIterable {
             case taskGroupId = "task_group_id"
             case cargoId = "cargo_id"
-            case driverId = "driver_id"
             case `public`
         }
     }