+++ Sources/Schemas/CargoByDriverSharingSettings1.swift
@@ -4,7 +4,7 @@ public struct CargoByDriverSharingSettings1: Codable, Hashable, Sendable {
     public let schemaVersion: Int
     public let taskGroupId: String
     public let cargoId: String
-    public let driverId: String
+    public let driverId: String?
     public let `public`: Bool?
     /// Must be a string starting with `user_`
     public let lastChangedByUserId: String?
@@ -19,7 +19,7 @@ public struct CargoByDriverSharingSettings1: Codable, Hashable, Sendable {
         schemaVersion: Int,
         taskGroupId: String,
         cargoId: String,
-        driverId: String,
+        driverId: String? = nil,
         public: Bool? = nil,
         lastChangedByUserId: String? = nil,
         lastChangedByOrgId: String? = nil,
@@ -44,7 +44,7 @@ public struct CargoByDriverSharingSettings1: Codable, Hashable, Sendable {
         self.schemaVersion = try container.decode(Int.self, forKey: .schemaVersion)
         self.taskGroupId = try container.decode(String.self, forKey: .taskGroupId)
         self.cargoId = try container.decode(String.self, forKey: .cargoId)
-        self.driverId = try container.decode(String.self, forKey: .driverId)
+        self.driverId = try container.decodeIfPresent(String.self, forKey: .driverId)
         self.public = try container.decodeIfPresent(Bool.self, forKey: .public)
         self.lastChangedByUserId = try container.decodeIfPresent(String.self, forKey: .lastChangedByUserId)
         self.lastChangedByOrgId = try container.decodeIfPresent(String.self, forKey: .lastChangedByOrgId)
@@ -59,7 +59,7 @@ public struct CargoByDriverSharingSettings1: Codable, Hashable, Sendable {
         try container.encode(self.schemaVersion, forKey: .schemaVersion)
         try container.encode(self.taskGroupId, forKey: .taskGroupId)
         try container.encode(self.cargoId, forKey: .cargoId)
-        try container.encode(self.driverId, forKey: .driverId)
+        try container.encodeIfPresent(self.driverId, forKey: .driverId)
         try container.encodeIfPresent(self.public, forKey: .public)
         try container.encodeIfPresent(self.lastChangedByUserId, forKey: .lastChangedByUserId)
         try container.encodeIfPresent(self.lastChangedByOrgId, forKey: .lastChangedByOrgId)